School System
The school system is a system that enables teachers and students to see their course and have the features of adding, deleting, and.  Update, we have three tables: the Teacher, the student and the course, for every Teacher he has: id and name and salary, for every student he has: id and name and age and Password , for every course he has: id and TeacherName and EnrolledStudent.